backend/requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
distlib==0.4.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pipenv==2025.0.4
platformdirs==4.3.8
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
SQLAlchemy==2.0.43
stripe==12.4.0
typing_extensions==4.14.1
urllib3==2.5.0
virtualenv==20.32.0
Werkzeug==3.1.3
